# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Capacitarr API
  version: 1.0.0
  description: |
    Capacitarr is a capacity management tool for media servers. It monitors disk
    usage across Sonarr, Radarr, Lidarr, Readarr, Plex, Jellyfin, Emby, Tautulli,
    and Overseerr integrations, scores media items for potential cleanup based on
    configurable weights and rules, and can automatically (or via approval) free
    disk space when thresholds are breached.

    ## Authentication

    Most endpoints require authentication via one of:
    - **JWT cookie** (`jwt`) — set automatically on login
    - **Bearer token** — `Authorization: Bearer <token>` header
    - **API key** — `X-Api-Key: <key>` header

    The `/health`, `/version`, and `/auth/login` endpoints are public.
  license:
    name: PolyForm Noncommercial 1.0.0
    url: https://polyformproject.org/licenses/noncommercial/1.0.0/
servers:
  - url: http://localhost:2187/api/v1
    description: Local development server
security:
  - cookieAuth: []
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /engine/history:
    get:
      operationId: getEngineHistory
      tags: [Engine]
      summary: Get engine run history
      description: |
        Returns historical engine run statistics within the specified time range.
        Each point represents one engine evaluation cycle with metrics for
        evaluated, flagged, deleted counts, freed bytes, and duration.
      parameters:
        - name: range
          in: query
          description: |
            Time range for history. Supports shorthand like `1h`, `6h`, `24h`,
            `7d`, `30d`.
          schema:
            type: string
            default: "7d"
            example: "7d"
      responses:
        "200":
          description: Array of engine history data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EngineHistoryPoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ===========================================================================
  # Integrations
  # ===========================================================================
...
  /audit:
    get:
      operationId: listAuditLogs
      tags: [Audit]
      summary: List audit logs
      description: |
        Returns paginated audit logs with optional filtering and sorting.
        The default sort is by `created_at` descending (newest first).
      parameters:
        - name: limit
          in: query
          description: Maximum number of records to return (max 1000)
          schema:
            type: integer
            default: 50
            maximum: 1000
        - name: offset
          in: query
          description: Number of records to skip for pagination
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          description: Filter by media name (substring match)
          schema:
            type: string
        - name: action
          in: query
          description: Filter by action type
          schema:
            type: string
            enum: [Deleted, Dry-Run]
        - name: sort_by
          in: query
          description: Column to sort by
          schema:
            type: string
            enum: [created_at, media_name, size_bytes, action]
            default: created_at
        - name: sort_dir
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated audit log response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /audit/grouped:
    get:
      operationId: getAuditGrouped
      tags: [Audit]
      summary: Get grouped audit logs
      description: |
        Returns audit logs grouped by TV show title for season/episode entries.
        Non-TV entries are returned as standalone items. Useful for a tree-style
        audit view.
      parameters:
        - name: limit
          in: query
          description: Maximum number of raw records to fetch before grouping (max 2000)
          schema:
            type: integer
            default: 200
            maximum: 2000
      responses:
        "200":
          description: Grouped audit results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupedAuditResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ===========================================================================
  # Data
  # ===========================================================================
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key generated via `POST /auth/apikey`
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login response
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: HttpOnly JWT cookie set on login

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
...
  responses:
    BadRequest:
      description: Bad request — invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required: [error]
      example:
        error: "Invalid request body"
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
          example: 1
        mediaName:
          type: string
          example: "Breaking Bad - Season 1"
        mediaType:
          type: string
          example: season
        reason:
          type: string
          description: Human-readable score summary (legacy format)
          example: "Score: 0.85 (WatchHistory: 1.0, Size: 0.5)"
        scoreDetails:
          type: string
          description: JSON-encoded array of ScoreFactor objects
          example: '[{"name":"Watch History","rawScore":1.0,"weight":10,"contribution":0.4,"type":"weight"}]'
        action:
          type: string
          enum: [Deleted, Dry-Run]
          example: Dry-Run
        sizeBytes:
          type: integer
          format: int64
          example: 5368709120
        createdAt:
          type: string
          format: date-time
      required: [id, mediaName, mediaType, reason, action, sizeBytes, createdAt]
    AuditResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
          description: Total number of records matching the filter
          example: 150
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
      required: [data, total, limit, offset]
    EngineHistoryPoint:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the engine run occurred
          example: "2026-03-01T14:00:00Z"
        evaluated:
          type: integer
          description: Number of media items evaluated
          example: 120
        flagged:
          type: integer
          description: Number of items flagged for deletion
          example: 8
        deleted:
          type: integer
          description: Number of items actually deleted
          example: 5
        freedBytes:
          type: integer
          format: int64
          description: Total bytes freed by deletions
          example: 2000000000
        durationMs:
          type: integer
          format: int64
          description: Engine run duration in milliseconds
          example: 320
      required: [timestamp, evaluated, flagged, deleted, freedBytes, durationMs]
    GroupedAuditResult:
      type: object
      properties:
        type:
          type: string
          enum: [group, single]
          description: Whether this is a grouped TV entry or a standalone item
        group:
          $ref: "#/components/schemas/AuditGroup"
        entry:
          $ref: "#/components/schemas/AuditLog"
    AuditGroup:
      type: object
      description: A group of audit entries belonging to the same TV show
      properties:
        showTitle:
          type: string
          example: "Breaking Bad"
        children:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        totalSize:
          type: integer
          format: int64
          description: Combined size of all children
        action:
          type: string
          description: Action of the first child (representative)
        createdAt:
          type: string
          format: date-time
          description: Timestamp of the first child (representative)
      required: [showTitle, children, totalSize, action, createdAt]

    # ---- Data Reset ----
...